backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
starlette==1.0.0
httpx==0.28.1
python-dotenv==1.1.0
